requirements.txt
requests == 2.28.2
python-telegram-bot == 13.15
python-dotenv == 1.0.0
